[Store] Split buffer pool bindings out of store_py.cpp
Move the registered buffer pool binding implementation into a dedicated
buffer_pool.cpp translation unit so the store module keeps only the
module wiring in store_py.cpp.
